+++ packages/las-sdk-core/src/types.ts
@@ -20,8 +20,9 @@ export type ListTransitionOptions = PaginationOptions & {
 }
 
 export interface CreateDocumentOptions {
-  consentId?: string;
   batchId?: string;
+  consentId?: string;
+  datasetId?: string;
   groundTruth?: Array<GroundTruth>;
 }
 
@@ -32,6 +33,7 @@ export interface UpdateDocumentOptions {
 export type ListDocumentsOptions = PaginationOptions & {
   batchId?: string | Array<string>;
   consentId?: string | Array<string>;
+  datasetId?: string | Array<string>;
 }
 
 export type TransitionExecutionList = {
@@ -210,6 +212,7 @@ export type GroundTruth = {
 export type DeleteDocumentOptions = PaginationOptions & {
   batchId?: string | Array<string>;
   consentId?: string | Array<string>;
+  datasetId?: string | Array<string>;
 }
 
 export type PostPredictions = CreatePredictionsOptions & {
@@ -249,12 +252,83 @@ export type Batch = {
   storageLocation: 'EU';
 };
 
+export type CreateBatchOptions = {
+  name?: string;
+  description?: string;
+  containsPersonallyIdentifiableInformation?: boolean;
+}
+
+export type UpdateBatchOptions = {
+  description?: string;
+  name?: string;
+}
+
 export type BatchList = {
   batches: Array<Batch>;
   nextToken: string | null;
 };
 
 export type ListBatchesOptions = PaginationOptions;
+
+export type Dataset = {
+  containsPersonallyIdentifiableInformation: boolean;
+  createdTime: string;
+  datasetId: string;
+  description: string;
+  name: string;
+  numDocuments: number;
+  retentionInDays: number;
+  storageLocation: 'EU';
+  updatedTime: string;
+  version: number;
+};
+
+export type CreateDatasetOptions = {
+  name?: string;
+  description?: string;
+  containsPersonallyIdentifiableInformation?: boolean;
+}
+
+export type UpdateDatasetOptions = {
+  description?: string;
+  name?: string;
+}
+
+export type DatasetList = {
+  datasets: Array<Dataset>;
+  nextToken: string | null;
+};
+
+export type ListDatasetsOptions = PaginationOptions;
+
+export type DataBundle = {
+  createdTime: string;
+  dataBundleId: string;
+  datasets: Array<Dataset>;
+  description: string;
+  modelId: string;
+  name: string;
+  status: 'ready' | 'processing' | 'failed';
+  summary: Record<string, any>;
+  updatedTime: string;
+};
+
+export type CreateDataBundleOptions = {
+  name?: string;
+  description?: string;
+}
+
+export type UpdateDataBundleOptions = {
+  description?: string;
+  name?: string;
+}
+
+export type DataBundleList = {
+  dataBundles: Array<DataBundle>;
+  nextToken: string | null;
+};
+
+export type ListDataBundleOptions = PaginationOptions;
 
 export type User = {
   userId: string;
@@ -306,16 +380,13 @@ export interface UpdateSecretOptions {
 export type LasDocumentWithoutContent = Omit<LasDocument, 'content'>;
 
 export type LasDocument = {
+  batchId?: string;
+  consentId?: string;
+  content: string;
   contentType: ContentType;
-  /** pattern: ^las:document:[a-f0-9]{32}$ */
+  datasetId?: string;
   documentId: string;
   groundTruth?: Array<GroundTruth>;
-  /** pattern: ^las:consent:[a-f0-9]{32}$ */
-  consentId?: string;
-  /** pattern: ^las:batch:[a-f0-9]{32}$ */
-  batchId?: string;
-  /** minimum: 1 */
-  content: string;
 };
 
 export type UpdateOrganizationOptions = {
@@ -325,6 +396,8 @@ export type UpdateOrganizationOptions = {
 
 export type Organization = {
   description: string | null;
+  monthlyNumberOfDataBundlesAllowed: number;
+  monthlyNumberOfDataBundlesCreated: number;
   monthlyNumberOfDocumentsAllowed: number;
   monthlyNumberOfDocumentsCreated: number;
   monthlyNumberOfPredictionsAllowed: number;
@@ -341,6 +414,8 @@ export type Organization = {
   numberOfAssetsCreated: number;
   numberOfBatchesAllowed: number;
   numberOfBatchesCreated: number;
+  numberOfDatasetsAllowed: number;
+  numberOfDatasetsCreated: number;
   numberOfModelsAllowed: number;
   numberOfModelsCreated: number;
   numberOfSecretsAllowed: number;
@@ -365,7 +440,9 @@ export type CreateAppClientOptions = {
 }
 
 export type UpdateAppClientOptions = {
+  defaultLoginUrl?: string;
   description?: string;
+  loginUrls?: Array<string>;
   name?: string;
 }
 
@@ -451,17 +528,6 @@ export type ListModelsOptions = PaginationOptions;
 export type ModelList = {
   models: Array<Model>;
   nextToken: string | null;
-}
-
-export type CreateBatchOptions = {
-  name?: string;
-  description?: string;
-  containsPersonallyIdentifiableInformation?: boolean;
-}
-
-export type UpdateBatchOptions = {
-  description?: string;
-  name?: string;
 }
 
 export interface UpdateAssetOptions {