+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog 
 
+## Version 4.4.0 - 2021-06-25
+
+- Added createDataBundle
+- Added updateDataBundle
+- Added deleteDataBundle
+- Added listDataBundles
+- Added defaultLoginUrl and loginUrls to updateAppClient options
+- Added monthlyNumberOfDataBundlesAllowed, monthlyNumberOfDataBundlesCreated, numberOfDatasetsAllowed and
+numberOfDatasetsCreated to Organization
+- Added deleteDocument
+- Added createDataset
+- Added updateDataset
+- Added deleteDataset
+- Added listDatasets
+- Deprecated createBatch
+- Deprecated updateBatch
+- Deprecated deleteBatch
+- Deprecated listBatches
+
 ## Version 4.3.1 - 2021-06-17
 
 - Fixed return type of createDocument